fix: false positives in recipe variants
