+++ .changeset/wild-snakes-give.md
@@ -1,0 +1,5 @@
+---
+'@pandacss/eslint-plugin': patch
+---
+
+Fix false positives in recipe variants
