+++ plugin/src/rules/index.ts
@@ -1,6 +1,7 @@
 import fileNotIncluded, { RULE_NAME as FileNotIncluded } from './file-not-included'
 import noConfigunctionInSource, { RULE_NAME as NoConfigunctionInSource } from './no-config-function-in-source'
 import noDebug, { RULE_NAME as NoDebug } from './no-debug'
+import noDeprecatedTokens, { RULE_NAME as NoDeprecatedTokens } from './no-deprecated-tokens'
 import noDynamicStyling, { RULE_NAME as NoDynamicStyling } from './no-dynamic-styling'
 import noEscapeHatch, { RULE_NAME as NoEscapeHatch } from './no-escape-hatch'
 import noHardCodedColor, { RULE_NAME as NoHardCodedColor } from './no-hardcoded-color'
@@ -21,6 +22,7 @@ export const rules = {
   [FileNotIncluded]: fileNotIncluded,
   [NoConfigunctionInSource]: noConfigunctionInSource,
   [NoDebug]: noDebug,
+  [NoDeprecatedTokens]: noDeprecatedTokens,
   [NoDynamicStyling]: noDynamicStyling,
   [NoEscapeHatch]: noEscapeHatch,
   [NoHardCodedColor]: noHardCodedColor,
