+++ plugin/src/utils/helpers.ts
@@ -283,3 +283,34 @@ export const getTaggedTemplateCaller = (node: TSESTree.TaggedTemplateExpression)
     return node.tag.callee.name
   }
 }
+
+export function isRecipeVariant(node: TSESTree.Property, context: RuleContext<any, any>) {
+  const caller = isInPandaFunction(node, context)
+  if (!caller) return
+
+  // Check if the caller is either 'cva' or 'sva'
+  const recipe = getImports(context).find((imp) => ['cva', 'sva'].includes(imp.name) && imp.alias === caller)
+  if (!recipe) return
+
+  //* Nesting is different here because of slots and variants. We don't want to warn about those.
+  let currentNode: any = node
+  let length = 0
+  let styleObjectParent = null
+
+  // Traverse up the AST
+  while (currentNode) {
+    if (currentNode.key && ['base', 'variants'].includes(currentNode.key.name)) {
+      styleObjectParent = currentNode.key.name
+    }
+    currentNode = currentNode.parent
+    if (!styleObjectParent) length++
+  }
+
+  // Determine the required length based on caller and styleObjectParent
+  const requiredLength = caller === 'cva' ? 2 : 4
+  const extraLength = styleObjectParent === 'base' ? 0 : 4
+
+  if (length >= requiredLength + extraLength) return
+
+  return true
+}
